+++ frontend/src/components/MemberStatus.vue
