+++ frontend/src/components/NewStory.vue
