+++ frontend/src/components/Nickname.vue
