+++ frontend/src/components/Observer.vue
