+++ frontend/src/components/Room.vue
