+++ frontend/src/components/StoryHistory.vue
