+++ frontend/src/components/StoryName.vue
